+++ batik-css/src/main/java/org/apache/batik/css/engine/StyleRule.java
@@ -31,7 +31,7 @@ public class StyleRule implements Rule {
     /**
      * The type constant.
      */
-    public static final short TYPE = (short)0;
+    public static final short TYPE = 0;
 
     /**
      * The selector list.
